Fixed some typos, code formatting, words, and sentence structure.
There were a number of typos (e.g. "python" instead of "Python" or
"it's" instead of "its"), words which were not formatted as code when
they should have been (to be consistent with other parts of the book),
misplaced words ("On a side now, [...]" I assumed should be "On a side
note, [...]"), and a couple of the sentences sounded weird. All these
have been changed/fixed.