Updated `cache_config.rst` to Python3.
The example of parameterisation at the end of the guide states that
`optparse` is deprecated and that `argparse` should be used instead.
However, it then proceeded to still use `optparse`. This has now been
changed. Additionally, the default values can be passed more elegantly
using `argparse` instead of the `if` statements in the constructors and
the guide has been updated to reflect this. Given that Python2 is
officially deprecated as of the 1st of January 2020, these changes
should bring the guide up to date to reflect that.